adjuster-test/pyproject.toml
[project]
name = "adjuster-test"
version = "0.1.0"
description = "Standalone harness to evaluate DejaQ context adjuster configs (tone matching)."
requires-python = ">=3.13"
dependencies = [
    "llama-cpp-python>=0.3.20",
    "sentence-transformers>=2.2.0",
    "numpy>=1.26.0",
    "tqdm>=4.65.0",
    "accelerate>=1.13.0",
    "anthropic>=0.52.0",
]

[tool.uv]
package = false
